Work around SELinux issues when uploading reports
Executing the script directly causes SELinux issues because the
dynflow workers running in foreman_rails_t are not allowed to execute
things of type foreman_lib_t. By wrapping the execution in a shell we
work around this we effectively work around the original issue because
foreman_rails_t is apparently allowed to execute bin_t.