Use hosts taxonomy when enabling cloud connector
Given the host that represents Foreman can only exist in one organization,
if a user has set a different taxonomy on the uploads page the
cloud connector will fail with a 404. The code needs to use the
taxonomies of the host when running cloud connector to ensure that it runs
no matter the context the user is in.