BZ2069640 - prevent hits#show network errors
When accessing the overview tab on the new host page,
there are two network error responses with a 500 status.
the first one which tries the endpoint `/insights_cloud/hits/3` is because that specific host doesn't have recommendations.

and the second which comes right after that is failing due to the path ends with `/insights_cloud/hits/undefined`

this doesn't cause errors to be seen on the UI,
but rather an API call that could be avoided.